+++ R/params-functions.R
@@ -15,7 +15,10 @@ getParams <- function(params, names) {
     checkmate::assertClass(params, classes = "Params")
     checkmate::assertCharacter(names, min.len = 1, any.missing = FALSE)
 
-    sapply(names, getParam, object = params, simplify = FALSE)
+    params.list <- lapply(names, getParam, object = params)
+    names(params.list) <- names
+
+    return(params.list)
 }
 
 #' Set parameters UNCHECKED
@@ -70,13 +73,18 @@ showPP <- function(params, pp) {
     for (category in names(pp)) {
         parameters <- pp[[category]]
         values <- getParams(params, parameters)
-        is.df <- sapply(values, is.data.frame)
+        is.df <- vapply(values, is.data.frame, FALSE)
 
         default.values <- getParams(default, parameters)
-        not.default <- sapply(seq_along(values), function(i) {
+        not.default <- vapply(seq_along(values), function(i) {
             !identical(values[i], default.values[i])
-        })
+        }, FALSE)
+        empty.values <- vapply(values, function(x) {
+            is.null(x) || length(x) == 0
+        }, FALSE)
+        values[empty.values] <- "Not set"
 
+        names(values) <- names(parameters)
         cat(crayon::bold(category), "\n")
         if (sum(!is.df) > 0) {
             showValues(values[!is.df], not.default[!is.df])
@@ -88,12 +96,14 @@ showPP <- function(params, pp) {
     }
 }
 
-#' Show vales
+#' Show values
 #'
-#' Function used for pretty printing scale or vector parameters.
+#' Function used for pretty printing scalar or vector parameters.
 #'
 #' @param values list of values to show.
 #' @param not.default logical vector giving which have changed from the default.
+#'
+#' @return Print values
 #'
 #' @importFrom utils head
 showValues <- function(values, not.default) {
@@ -102,13 +112,22 @@ showValues <- function(values, not.default) {
     checkmate::check_logical(not.default, any.missing = FALSE,
                              len = length(values))
 
-    short.values <- sapply(values, function(x) {
-        if (length(x) > 4) {
-            paste0(paste(head(x, n = 4), collapse = ", "), ",...")
+    short.values <- vapply(values, function(x) {
+        if (is.list(x)) {
+            classes <- class(x)
+            if (length(classes) == 1 && classes == "list") {
+                paste("List with", length(x), "items")
+            } else {
+                paste("Object of class", paste(classes, collapse = ", "))
+            }
         } else {
-            paste(x, collapse = ", ")
+            if (length(x) > 4) {
+                paste0(paste(head(x, n = 4), collapse = ", "), ",...")
+            } else {
+                paste(x, collapse = ", ")
+            }
         }
-    })
+    }, c(Value = "None"))
 
     names(short.values)[not.default] <- toupper(names(values[not.default]))
 
@@ -118,14 +137,18 @@ showValues <- function(values, not.default) {
     items.per.line <- floor(screen.width / (max.len + 2))
 
     short.names <- names(short.values)
-    short.values <- crayon::col_align(short.values, max.len, "right")
-    short.names <- crayon::col_align(short.names, max.len, "right")
-
     not.est <- !grepl("\\(", short.names)
+    secondary <- grepl("\\*", short.names)
+    short.names <- gsub("\\*", "", short.names)
+
     short.names[not.est] <- crayon::blue(short.names[not.est])
+    short.names[secondary] <- crayon::bgYellow(short.names[secondary])
     short.names[not.default] <- crayon::bold(short.names[not.default])
     short.values[not.default] <- crayon::green(short.values[not.default])
     short.values[not.default] <- crayon::bold(short.values[not.default])
+
+    short.values <- crayon::col_align(short.values, max.len, "right")
+    short.names <- crayon::col_align(short.names, max.len, "right")
 
     names(short.values) <- short.names
 
@@ -144,6 +167,8 @@ showValues <- function(values, not.default) {
 #'
 #' @param dfs list of data.frames to show.
 #' @param not.default logical vector giving which have changed from the default.
+#'
+#' @return Print data.frame parameters
 #'
 #' @importFrom utils head
 showDFs <- function(dfs, not.default) {
@@ -173,6 +198,8 @@ showDFs <- function(dfs, not.default) {
         cat(paste0("\n", name, "\n"))
         cat(msg, "\n")
         print(head(df, n = 4))
-        cat("# ... with", nrow(df) - 4, "more rows\n")
+        if (nrow(df) > 4) {
+            cat("# ... with", nrow(df) - 4, "more rows\n")
+        }
     }
 }